+++ src/pages/api/server/routes/user.router.ts
@@ -1,5 +1,6 @@
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime";
 import { TRPCError } from "@trpc/server";
+import { unstable_getServerSession } from "next-auth";
 import {
   inviteUserInput,
   inviteUserOutput,
@@ -12,7 +13,6 @@ import {
   userDeleteInput,
   UserDeleteOutput,
   userDeleteOutput,
-  userDetailsInput,
   UserDetailsOutput,
   userDetailsOutput,
   userListInput,
@@ -23,6 +23,7 @@ import {
   userRoleOutput,
   userSearchInput,
 } from "../../../../schema/user.schema";
+import { authOptions } from "../../auth/[...nextauth]";
 import { createRouter } from "../createRouter";
 import { handlePrismaError } from "../errors/prisma.errors";
 import { NodemailerInstance } from "../nodemailer_instance";
@@ -39,6 +40,14 @@ export const userRouter = createRouter()
       };
 
       try {
+                // only admin is allowed to invite users
+                const session = await unstable_getServerSession(
+                  ctx?.req!,
+                  ctx?.res!,
+                  authOptions
+                );
+        
+                if (session?.user.userStatus !== "INACTIVE") {
         const updatedDBUser = await ctx?.prisma.user.update({
           where: {
             email: input.email,
@@ -91,6 +100,13 @@ export const userRouter = createRouter()
             },
           });
         }
+      } else {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          cause: "Improper authorization",
+          message: "inactive users are not allowed to make the request",
+        });
+      }
       } catch (err) {
         if (err instanceof PrismaClientKnownRequestError)
           handlePrismaError(err);
@@ -105,39 +121,61 @@ export const userRouter = createRouter()
     output: inviteUserOutput,
     async resolve({ ctx, input }) {
       try {
-        if (!input.email.endsWith("@aitpune.edu.in")) {
-          throw new TRPCError({
-            code: "BAD_REQUEST",
-            cause: "invalid email id",
-            message:
-              "only user with email having @aitpune.edu.in can be invited",
+        // only admin is allowed to invite users
+        const session = await unstable_getServerSession(
+          ctx?.req!,
+          ctx?.res!,
+          authOptions
+        );
+
+        if (
+          (session?.user.role == "ADMIN" ||
+            session?.user.role == "SUPER_ADMIN") &&
+          session?.user.userStatus == "ACTIVE"
+        ) {
+          if (!input.email.endsWith("@aitpune.edu.in")) {
+            throw new TRPCError({
+              code: "BAD_REQUEST",
+              cause: "invalid email id",
+              message:
+                "only user with email having @aitpune.edu.in can be invited",
+            });
+          }
+          const user = await ctx?.prisma.user.create({
+            data: {
+              email: input.email,
+              role: input.role,
+              userStatus: "INVITED",
+            },
+            select: {
+              email: true,
+              role: true,
+            },
           });
-        }
-        const user = await ctx?.prisma.user.create({
-          data: {
-            email: input.email,
-            role: input.role,
-            userStatus: "INVITED",
-          },
-          select: {
-            email: true,
-            role: true,
-          },
-        });
-        if (user) {
-          // nodemailer
-          await NodemailerInstance.GetNodemailer();
-          await NodemailerInstance.SendUserInviteEmail(input.email, input.role);
-          return {
-            email: user.email,
-            role: user.role as ROLES,
-          };
+          if (user) {
+            // nodemailer
+            await NodemailerInstance.GetNodemailer();
+            await NodemailerInstance.SendUserInviteEmail(
+              input.email,
+              input.role
+            );
+            return {
+              email: user.email,
+              role: user.role as ROLES,
+            };
+          }
+        } else {
+          throw new TRPCError({
+            code: "UNAUTHORIZED",
+            cause: "Improper authorization",
+            message:
+              "only admins and super admins are allowed to make the request",
+          });
         }
       } catch (e) {
         if (e instanceof PrismaClientKnownRequestError) handlePrismaError(e);
         else console.log(e);
       }
-
       // return fail response
       return {
         email: "",
@@ -154,34 +192,46 @@ export const userRouter = createRouter()
           email: "",
           id: "",
           name: "",
-          phoneNo: "",
           role: "STUDENT",
           userStatus: "INACTIVE",
         },
       ];
       try {
-        const dbStudent = await ctx?.prisma.user.findMany({
-          where: {
-            role: input.role as ROLES,
-          },
-          select: {
-            id: true,
-            email: true,
-            name: true,
-            role: true,
-            phoneNo: true,
-            userStatus: true,
-          },
-        });
+        // only admin is allowed to invite users
+        const session = await unstable_getServerSession(
+          ctx?.req!,
+          ctx?.res!,
+          authOptions
+        );
 
-        response = dbStudent?.map((item) => ({
-          id: item.id,
-          email: item.email,
-          name: item.name || "N.A",
-          role: item.role as ROLES,
-          phoneNo: item.phoneNo || "N.A",
-          userStatus: item.userStatus as USER_STATUS,
-        }))!;
+        if (session?.user.userStatus == "ACTIVE") {
+          const dbStudent = await ctx?.prisma.user.findMany({
+            where: {
+              role: input.role as ROLES,
+            },
+            select: {
+              id: true,
+              email: true,
+              name: true,
+              role: true,
+              userStatus: true,
+            },
+          });
+
+          response = dbStudent?.map((item) => ({
+            id: item.id,
+            email: item.email,
+            name: item.name || "N.A",
+            role: item.role as ROLES,
+            userStatus: item.userStatus as USER_STATUS,
+          }))!;
+        } else {
+          throw new TRPCError({
+            code: "UNAUTHORIZED",
+            cause: "Improper authorization",
+            message: "only active users are allowed to make the request",
+          });
+        }
       } catch (err) {
         if (err instanceof PrismaClientKnownRequestError)
           handlePrismaError(err);
@@ -190,10 +240,9 @@ export const userRouter = createRouter()
       return response;
     },
   })
-  .query("get-user-details", {
-    input: userDetailsInput,
+  .query("get-user-profile-details", {
     output: userDetailsOutput,
-    async resolve({ ctx, input }) {
+    async resolve({ ctx }) {
       let response: UserDetailsOutput = {
         email: "",
         name: "",
@@ -202,38 +251,53 @@ export const userRouter = createRouter()
         userStatus: "INACTIVE",
       };
       try {
-        const dbUser = await ctx?.prisma.user.findFirst({
-          where: {
-            email: input.email,
-          },
-          select: {
-            email: true,
-            name: true,
-            role: true,
-            phoneNo: true,
-            userStatus: true,
-            StudentDetails: {
-              select: {
-                branch: true,
-                registrationNumber: true,
-                year: true,
+        // only admin is allowed to invite users
+        const session = await unstable_getServerSession(
+          ctx?.req!,
+          ctx?.res!,
+          authOptions
+        );
+
+        if (session?.user.userStatus == "ACTIVE") {
+          const dbUser = await ctx?.prisma.user.findFirst({
+            where: {
+              email: session.user.email,
+            },
+            select: {
+              email: true,
+              name: true,
+              role: true,
+              phoneNo: true,
+              userStatus: true,
+              StudentDetails: {
+                select: {
+                  branch: true,
+                  registrationNumber: true,
+                  year: true,
+                },
               },
             },
-          },
-        });
-        response = {
-          email: dbUser?.email!,
-          name: dbUser?.name || "N.A",
-          role: dbUser?.role as ROLES,
-          phoneNo: dbUser?.phoneNo || "N.A",
-          userStatus: dbUser?.userStatus as USER_STATUS,
-        };
-        if (dbUser?.StudentDetails) {
-          response.studentDetails = {
-            branch: dbUser.StudentDetails.branch!,
-            registrationNumber: dbUser.StudentDetails.registrationNumber,
-            year: dbUser.StudentDetails.year!,
+          });
+          response = {
+            email: dbUser?.email!,
+            name: dbUser?.name || "N.A",
+            role: dbUser?.role as ROLES,
+            phoneNo: dbUser?.phoneNo || "N.A",
+            userStatus: dbUser?.userStatus as USER_STATUS,
           };
+          if (dbUser?.StudentDetails) {
+            response.studentDetails = {
+              branch: dbUser.StudentDetails.branch!,
+              registrationNumber: dbUser.StudentDetails.registrationNumber,
+              year: dbUser.StudentDetails.year!,
+            };
+          }
+        } else {
+          throw new TRPCError({
+            code: "UNAUTHORIZED",
+            cause: "Improper authorization",
+            message: "only active users are allowed to make the request",
+          });
         }
       } catch (err) {
         if (err instanceof PrismaClientKnownRequestError)
@@ -252,19 +316,39 @@ export const userRouter = createRouter()
         role: "STUDENT",
       };
       try {
-        const dbUser = await ctx?.prisma.user.update({
-          where: {
-            id: input.id,
-          },
-          data: {
-            role: input.role,
-          },
-        })!;
+        // only admin is allowed to invite users
+        const session = await unstable_getServerSession(
+          ctx?.req!,
+          ctx?.res!,
+          authOptions
+        );
 
-        response = {
-          email: dbUser.email,
-          role: dbUser.role,
-        };
+        if (
+          (session?.user.role == "ADMIN" ||
+            session?.user.role == "SUPER_ADMIN") &&
+          session?.user.userStatus == "ACTIVE"
+        ) {
+          const dbUser = await ctx?.prisma.user.update({
+            where: {
+              id: input.id,
+            },
+            data: {
+              role: input.role,
+            },
+          })!;
+
+          response = {
+            email: dbUser.email,
+            role: dbUser.role,
+          };
+        } else {
+          throw new TRPCError({
+            code: "UNAUTHORIZED",
+            cause: "Improper authorization",
+            message:
+              "only admins and super admins are allowed to make the request",
+          });
+        }
       } catch (err) {
         if (err instanceof PrismaClientKnownRequestError)
           handlePrismaError(err);
@@ -280,16 +364,37 @@ export const userRouter = createRouter()
       let response: UserDeleteOutput = {
         email: "",
       };
-      try {
-        const dbUser = await ctx?.prisma.user.delete({
-          where: {
-            id: input.id,
-          },
-        })!;
 
-        return {
-          email: dbUser.email,
-        };
+      try {
+        // only admin is allowed to invite users
+        const session = await unstable_getServerSession(
+          ctx?.req!,
+          ctx?.res!,
+          authOptions
+        );
+
+        if (
+          (session?.user.role == "ADMIN" ||
+            session?.user.role == "SUPER_ADMIN") &&
+          session?.user.userStatus == "ACTIVE"
+        ) {
+          const dbUser = await ctx?.prisma.user.delete({
+            where: {
+              id: input.id,
+            },
+          })!;
+
+          return {
+            email: dbUser.email,
+          };
+        } else {
+          throw new TRPCError({
+            code: "UNAUTHORIZED",
+            cause: "Improper authorization",
+            message:
+              "only admins and super admins are allowed to make the request",
+          });
+        }
       } catch (e) {
         if (e instanceof PrismaClientKnownRequestError) {
           handlePrismaError(e);
@@ -305,31 +410,46 @@ export const userRouter = createRouter()
     async resolve({ ctx, input }) {
       let response: UserListOutput = [];
       try {
-        const searchProcessedString = input.searchText
-          .replace(/[^a-zA-Z0-9 ]/g, "") // remove special charachters
-          .replaceAll(/\s/g, "") // remove multiple whitespace
-          .trim(); // remove starting and trailing spaces
-        //.replaceAll(" ", " | "); // add or
+        // only admin is allowed to invite users
+        const session = await unstable_getServerSession(
+          ctx?.req!,
+          ctx?.res!,
+          authOptions
+        );
 
-        const dbUserSearch = await ctx?.prisma.user.findMany({
-          where: {
-            email: {
-              contains: searchProcessedString,
-              mode: "insensitive",
+        if (session?.user.userStatus == "ACTIVE") {
+          const searchProcessedString = input.searchText
+            .replace(/[^a-zA-Z0-9 ]/g, "") // remove special charachters
+            .replaceAll(/\s/g, "") // remove multiple whitespace
+            .trim(); // remove starting and trailing spaces
+          //.replaceAll(" ", " | "); // add or
+
+          const dbUserSearch = await ctx?.prisma.user.findMany({
+            where: {
+              email: {
+                contains: searchProcessedString,
+                mode: "insensitive",
+              },
             },
-          },
-        });
+          });
 
-        response = dbUserSearch?.map((user) => {
-          return {
-            email: user.email,
-            id: user.id,
-            name: user.name || "N.A",
-            phoneNo: user.phoneNo || "N.A",
-            role: user.role,
-            userStatus: user.userStatus,
-          };
-        })!;
+          response = dbUserSearch?.map((user) => {
+            return {
+              email: user.email,
+              id: user.id,
+              name: user.name || "N.A",
+              phoneNo: user.phoneNo || "N.A",
+              role: user.role,
+              userStatus: user.userStatus,
+            };
+          })!;
+        } else {
+          throw new TRPCError({
+            code: "UNAUTHORIZED",
+            cause: "Improper authorization",
+            message: "only active accounts allowed to make the request",
+          });
+        }
       } catch (e) {
         if (e instanceof PrismaClientKnownRequestError) {
           handlePrismaError(e);
